[backend] move subtractInterval helper to common module (#2197)
